fix: move gutters definition to lib as it needs Theme avoiding cycle
