+++ src/lib/index.ts
@@ -1,2 +1,3 @@
 export * from './common.types';
+export * from './get-gutters';
 export * from './styled-components.config';
