--- src/theming/gutters/index.ts
@@ -1,1 +1,0 @@
-export * from './gutters';
