+++ src/theming/palette/info/info.ts
