--- src/theming/palettes/index.ts
@@ -1,8 +1,0 @@
-export * from './error/error';
-export * from './info/info';
-export * from './neutral/neutral';
-export * from './primary/primary';
-export * from './report/report';
-export * from './secondary/secondary';
-export * from './success/success';
-export * from './warning/warning';
