+++ src/theming/theme/theme.ts
@@ -1,37 +1,16 @@
-import { useTheme as useStyledTheme } from 'styled-components';
+import { defaultGutters, GutterThemeProps } from '../gutters/gutters';
+import { defaultPalette, PaletteThemeKeys } from '../palette/palette';
 
-import { DeepNestedKeyOf } from '../../lib';
-import {
-  error,
-  info,
-  neutral,
-  primary,
-  report,
-  secondary,
-  success,
-  warning,
-} from '../palettes';
+export type ThemeProps = {
+  palette: PaletteThemeKeys;
+  gutter?: GutterThemeProps;
+  roundness?: number | string;
+};
 
-export const defaultTheme = {
-  palette: {
-    primary,
-    secondary,
-    neutral,
-    info,
-    success,
-    warning,
-    error,
-    report,
-  },
-  gutter: {
-    type: 'em',
-    size: 0.25,
-  },
+export const defaultTheme: ThemeProps = {
+  palette: defaultPalette,
+  gutter: defaultGutters,
   roundness: 4,
 };
 
-export type DefaultTheme = typeof defaultTheme;
-export const useTheme = useStyledTheme as () => DefaultTheme;
-
-export type PaletteThemeKeys = DeepNestedKeyOf<DefaultTheme['palette']>;
-// export type GradienthemeKeys = DeepNestedKeyOf<DefaultTheme['gradients']>;
+export type Theme = typeof defaultTheme;
